Added All Component Structure for Article Page
I added the basic structure of article page.

Images and title are written manually, will be changed in future commit, where image, title, and date are displayed based on articles table in database.

Also added images used for articles to the public/img folders so that the database can save the path of the image.